refactor(proxy): move BYO/local-CA TLS decision into tls::build_proxy_tls
The four-arm match and its `build_local_ca_tls` helper both lived in
serve.rs even though every arm called into tls.rs. Fold the whole
decision — enabled-guard, BYO, half-set error, local-CA fallback — into
one function next to `load_pem_tls_config`. serve.rs's startup path
becomes a single line; TLS policy is co-located with the TLS code.

Also drop the redundant `key.display()` from the BYO load-failure warn —
`load_pem_tls_config` already names the offending file in its error.
